Moved CustomMatchers to matchers package
